`console` command now honors non-Default AWS SSO instances
Users using `--sso` with `eval` or `exec` and then running
`console` would find the latter command would fail because
it did not know which AWS SSO instance was used with the
previous command.

We now set AWS_SSO environment variable so this value is
automatically tracked with `console`, `list`, etc.

Fixes: #208