console command now supports $AWS_PROFILE
After all other existing checks, look to see if $AWS_PROFILE
is set and if so, use those credentials to open the console with.

Since not all profiles are necesssarily AWS SSO profiles, we will
try using the Go SDK if necessary to load our credentials instead
of the creds we have cached.

Fixes: #234